refactor(#577): move finish-counter
Moved finish-counter into its own file to reduce
the size of the cluster-handler.
