+++ src/binaries/hanami/src/core/blocks/core_block.rs
@@ -24,7 +24,7 @@ use ainari_common::enums::*;
 use ainari_common::error::AinariError;
 use ainari_common::functions::*;
 
-use crate::core::cluster_handler::FinishCounter;
+use crate::core::processing::finish_counter::FinishCounter;
 
 use super::axons::*;
 use super::block_io::*;
