+++ src/binaries/hanami/src/core/blocks/input_block.rs
@@ -21,7 +21,7 @@ use super::axons::*;
 use super::block_io::*;
 use super::block_trait::*;
 
-use crate::core::cluster_handler::*;
+use crate::core::processing::finish_counter::FinishCounter;
 
 use ainari_common::constants::*;
 use ainari_common::enums::*;
