+++ src/binaries/hanami/src/core/blocks/new_core_block.rs
@@ -20,7 +20,7 @@ use ainari_common::constants::*;
 use ainari_common::enums::*;
 use ainari_common::error::AinariError;
 
-use crate::core::cluster_handler::FinishCounter;
+use crate::core::processing::finish_counter::FinishCounter;
 
 use super::axons::*;
 use super::block_io::*;
