+++ src/binaries/hanami/src/core/blocks/output_block.rs
@@ -22,6 +22,7 @@ use ainari_common::enums::*;
 use ainari_common::error::AinariError;
 
 use crate::core::cluster_handler::*;
+use crate::core::processing::finish_counter::FinishCounter;
 use crate::core::processing::output_buffer::*;
 
 use super::axons::*;
