+++ src/binaries/hanami/src/core/cluster_handler.rs
@@ -29,69 +29,14 @@ use ainari_common::error::AinariError;
 use crate::core::blocks::core_block::*;
 use crate::core::blocks::input_block::*;
 use crate::core::blocks::output_block::*;
+use crate::core::processing::finish_counter::FinishCounter;
 use crate::core::processing::output_buffer::OutputBuffer;
 
 use super::blocks::block_trait::Block;
 use super::cluster_interface::ClusterInterface;
-use super::processing::tasks::Task;
 
 lazy_static::lazy_static! {
     pub static ref CLUSTER_HANDLER: RwLock<ClusterDataHandler> = RwLock::new(init_cluster_data_handler());
-}
-
-// ==================================================================================================
-
-#[derive(Default, Debug)]
-pub struct FinishCounter {
-    pub input_compare: usize,
-    pub output_compare: usize,
-    task_compare: usize,
-    counter: usize,
-
-    expected_cycle_number: u64,
-    already_finished: bool,
-
-    pub task: Option<Arc<Mutex<Task>>>,
-}
-
-impl FinishCounter {
-    pub fn reset(&mut self, task_compare: usize, expected_cycle_number: u64) {
-        self.expected_cycle_number = expected_cycle_number;
-        self.counter = 0;
-        self.already_finished = false;
-        self.task_compare = task_compare;
-        self.task = None;
-    }
-
-    pub fn update(&mut self, cycle_number: u64) {
-        if cycle_number == self.expected_cycle_number {
-            self.counter += 1;
-
-            if self.is_finished() && !self.already_finished {
-                self.already_finished = true;
-                if let Some(task_mutex) = &self.task {
-                    let mut task = task_mutex.lock().unwrap();
-                    task.finisch_cycle();
-                    let next_cycle = self.expected_cycle_number + 1;
-                    self.expected_cycle_number = next_cycle;
-                    self.counter = 0;
-                    self.already_finished = false;
-                }
-            }
-        }
-    }
-
-    pub fn is_finished(&self) -> bool {
-        if self.counter >= self.task_compare {
-            return true;
-        }
-
-        false
-    }
-
-    pub fn get_expected_cycle_number(&self) -> u64 {
-        self.expected_cycle_number
-    }
 }
 
 // ==================================================================================================
