+++ src/binaries/hanami/src/core/processing/mod.rs
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+pub mod finish_counter;
 pub mod output_buffer;
 pub mod task_queue;
 pub mod tasks;
